refactor: move file endpoints to hono
