+++ apps/web/package.json
@@ -10,7 +10,6 @@
     "dev": "vp dev"
   },
   "dependencies": {
-    "@aws-sdk/client-s3": "^3.1068.0",
     "@mantine/charts": "^9.3.1",
     "@mantine/core": "^9.3.1",
     "@mantine/dates": "^9.3.1",
@@ -18,10 +17,7 @@
     "@mantine/hooks": "^9.3.1",
     "@mantine/notifications": "^9.3.1",
     "@mantine/schedule": "^9.3.1",
-    "@paralleldrive/cuid2": "^3.3.0",
     "@prive-admin-tanstack/api": "workspace:*",
-    "@prive-admin-tanstack/auth": "workspace:*",
-    "@prive-admin-tanstack/db": "workspace:*",
     "@prive-admin-tanstack/env": "workspace:*",
     "@prive-admin-tanstack/ui": "workspace:*",
     "@tabler/icons-react": "^3.44.0",
@@ -30,11 +26,8 @@
     "@trpc/client": "catalog:",
     "@trpc/server": "catalog:",
     "@trpc/tanstack-react-query": "catalog:",
-    "archiver": "^8.0.0",
     "better-auth": "catalog:",
-    "csv-parse": "^6.2.1",
     "dayjs": "^1.11.21",
-    "drizzle-orm": "^0.45.2",
     "mantine-form-zod-resolver": "^1.3.0",
     "react": "catalog:",
     "react-dom": "catalog:",
