--- apps/web/src/functions/get-locale.ts
@@ -1,9 +1,0 @@
-import { createServerFn } from "@tanstack/react-start"
-
-import { localeMiddleware } from "@/middleware/locale"
-
-export const getLocale = createServerFn({ method: "GET" })
-  .middleware([localeMiddleware])
-  .handler(async ({ context }) => {
-    return { locale: context.locale, timeZone: context.timeZone }
-  })
