--- apps/web/src/functions/get-user.ts
@@ -1,9 +1,0 @@
-import { createServerFn } from "@tanstack/react-start"
-
-import { authMiddleware } from "@/middleware/auth"
-
-export const getUser = createServerFn({ method: "GET" })
-  .middleware([authMiddleware])
-  .handler(async ({ context }) => {
-    return context.session
-  })
