--- apps/web/src/functions/transactions.ts
@@ -1,112 +1,0 @@
-import { db } from "@prive-admin-tanstack/db"
-import { personnelOnAppointments } from "@prive-admin-tanstack/db/schema/appointment"
-import { transaction } from "@prive-admin-tanstack/db/schema/transaction"
-import { createServerFn } from "@tanstack/react-start"
-import { eq } from "drizzle-orm"
-import { z } from "zod"
-
-import { currencySchema } from "@/lib/currency"
-import { requireAuthMiddleware } from "@/middleware/auth"
-
-const transactionFieldsSchema = z.object({
-  name: z.string().nullish(),
-  notes: z.string().nullish(),
-  amount: z.number().int(),
-  currency: currencySchema,
-})
-
-export const getTransactionsByAppointmentId = createServerFn({ method: "GET" })
-  .middleware([requireAuthMiddleware])
-  .inputValidator(z.object({ appointmentId: z.string() }))
-  .handler(async ({ data }) => {
-    return db.query.transaction.findMany({
-      where: eq(transaction.appointmentId, data.appointmentId),
-      with: {
-        customer: { columns: { id: true, name: true } },
-      },
-      orderBy: (tx, { asc }) => [asc(tx.createdAt)],
-    })
-  })
-
-export const createTransaction = createServerFn({ method: "POST" })
-  .middleware([requireAuthMiddleware])
-  .inputValidator(
-    transactionFieldsSchema.extend({
-      appointmentId: z.string().min(1),
-      customerId: z.string().min(1),
-    }),
-  )
-  .handler(async ({ data }) => {
-    return await db.transaction(async (tx) => {
-      const allowedCustomerIds = new Set<string>()
-      const appointmentRow = await tx.query.appointment.findFirst({
-        where: (a, { eq }) => eq(a.id, data.appointmentId),
-        columns: { clientId: true },
-      })
-      if (!appointmentRow) {
-        throw new Error("Appointment not found")
-      }
-      allowedCustomerIds.add(appointmentRow.clientId)
-      const personnelRows = await tx
-        .select({ personnelId: personnelOnAppointments.personnelId })
-        .from(personnelOnAppointments)
-        .where(eq(personnelOnAppointments.appointmentId, data.appointmentId))
-      for (const row of personnelRows) {
-        allowedCustomerIds.add(row.personnelId)
-      }
-      if (!allowedCustomerIds.has(data.customerId)) {
-        throw new Error("Customer is not the appointment client or assigned personnel")
-      }
-
-      const [result] = await tx
-        .insert(transaction)
-        .values({
-          appointmentId: data.appointmentId,
-          customerId: data.customerId,
-          name: data.name ?? null,
-          notes: data.notes ?? null,
-          amount: data.amount,
-          currency: data.currency,
-        })
-        .returning()
-      return result
-    })
-  })
-
-export const updateTransaction = createServerFn({ method: "POST" })
-  .middleware([requireAuthMiddleware])
-  .inputValidator(transactionFieldsSchema.extend({ id: z.string().min(1) }))
-  .handler(async ({ data }) => {
-    const existing = await db.query.transaction.findFirst({
-      where: eq(transaction.id, data.id),
-      columns: { id: true },
-    })
-    if (!existing) {
-      throw new Error("Transaction not found")
-    }
-    const [result] = await db
-      .update(transaction)
-      .set({
-        name: data.name ?? null,
-        notes: data.notes ?? null,
-        amount: data.amount,
-        currency: data.currency,
-      })
-      .where(eq(transaction.id, data.id))
-      .returning()
-    return result
-  })
-
-export const deleteTransaction = createServerFn({ method: "POST" })
-  .middleware([requireAuthMiddleware])
-  .inputValidator(z.object({ id: z.string().min(1) }))
-  .handler(async ({ data }) => {
-    const existing = await db.query.transaction.findFirst({
-      where: eq(transaction.id, data.id),
-      columns: { id: true },
-    })
-    if (!existing) {
-      throw new Error("Transaction not found")
-    }
-    await db.delete(transaction).where(eq(transaction.id, data.id))
-  })
