--- apps/web/src/routes/api/auth/$.ts
@@ -1,15 +1,0 @@
-import { auth } from "@prive-admin-tanstack/auth"
-import { createFileRoute } from "@tanstack/react-router"
-
-export const Route = createFileRoute("/api/auth/$")({
-  server: {
-    handlers: {
-      GET: ({ request }) => {
-        return auth.handler(request)
-      },
-      POST: ({ request }) => {
-        return auth.handler(request)
-      },
-    },
-  },
-})
